py1/requirements_microservices.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.2